requirements.txt
fastapi
uvicorn
supabase
python-dotenv
PyJWT
